fix cli and gui to adapt to scope statement specs

+++ src/com/company/behavior/IAMonkeyStrategy.java
@@ -16,7 +16,7 @@ public class IAMonkeyStrategy implements Behavior {
         int lowerBound = 1;    //included
         int higherBound = this.systemWidth+1;   //excluded
 
-        return Double.toString( Math.random() * (higherBound-lowerBound) + lowerBound); //random choice
+        return Integer.toString( (int) (Math.random() * (higherBound-lowerBound) + lowerBound) ); //random choice
         //return (Double.toString(Math.random() * (7 - 1 + 1)) + 1);      //TODO adapt with width of grid, ok done, test it if ok?
     }
